requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-bcrypt
psycopg2-binary
python-dotenv
flask-cors
redis